refactor for widgets
